Map all endpoints of interfaces grouped by IAD to first interface in the association.
This is a workaround for a windows quirk whereby interfaces grouped by an Interface
Association Descriptor (IAD) on composite devices are lumped into a single PDO.  See:

https://msdn.microsoft.com/en-us/library/windows/hardware/ff537107(v=vs.85).aspx

Apparently, not a lot of devices use IADs, but one class of devices in particular that
are causing a lot of headaches are AIA U3V class cameras.  Those cameras create an
IAD that assocates three interfaces : a command channel, a event channel, and a stream
intput channel (usually in that order).

This patch maps all of the endpoints for the grouped interface to the first interface
of the group in the private libusb configuration structure.

The results:
- claiming the first interface of the IAD group is successful.
- claiming the subsequent interfaces in the IAD group will fail (harmlessly ?).
- submitting transfers to the endpoints on any of the interfaces will work, largely
  because the API locates the interface for the endpoints using the
  get_interface_by_endpoint() call, which is fooled by the endpoint mapping logic in this
  patch.

It "works in the lab", but it feels / looks like a hack.